requirements.txt
pdfminer.six
chardet
unidecode

# dev
pylint
ipython
black
pytest
